Refactor main.js: prefs controller
